createExampleTable -> ExampleTable
