+++ frontend/src/components/Analysis/AnalysisReport/index.tsx
@@ -4,17 +4,12 @@ import {
   ResultFineGrainedParsed,
   BucketIntervals,
 } from "../types";
-import { compareBucketOfCases } from "../utils";
-import { AnalysisTable } from "../../../components";
-import { Typography, Space, Tabs } from "antd";
 import { SystemModel } from "../../../models";
 import { SystemAnalysesReturn } from "../../../clients/openapi/api";
 import { OverallMetricsBarChart } from "./OverallMetricsBarChart";
 import { SignificanceTestList } from "./SignificanceTestList";
 import { FineGrainedAnalysis } from "./FineGrainedAnalysis";
-
-const { Title } = Typography;
-const { TabPane } = Tabs;
+import { ExampleTable } from "./ExampleTable";
 
 interface Props {
   task: string;
@@ -29,85 +24,6 @@ interface Props {
     featureName: string,
     bucketInfo: BucketIntervals
   ) => void;
-}
-
-function createExampleTable(
-  props: Props,
-  activeSystemExamples: ActiveSystemExamples | undefined,
-  setActiveSystemExamples: React.Dispatch<
-    React.SetStateAction<ActiveSystemExamples | undefined>
-  >,
-  page: number,
-  setPage: React.Dispatch<React.SetStateAction<number>>
-) {
-  // Create a table full of examples
-  let exampleTable = <div>&nbsp;</div>;
-  if (activeSystemExamples === undefined) {
-    return exampleTable;
-  }
-
-  const { task, systems } = props;
-  const { title, barIndex, systemIndex, bucketOfCasesList } =
-    activeSystemExamples;
-
-  // Sort bucket of samples for every system
-  const sortedBucketOfCasesList = bucketOfCasesList.map((bucketOfCases) => {
-    return bucketOfCases.sort(compareBucketOfCases);
-  });
-
-  // single analysis
-  if (systems.length === 1) {
-    exampleTable = (
-      <AnalysisTable
-        systemID={systems[0].system_id}
-        task={task}
-        cases={sortedBucketOfCasesList[0]}
-        page={page}
-        onPageChange={setPage}
-      />
-    );
-    // multi-system analysis
-  } else {
-    exampleTable = (
-      <Space style={{ width: "fit-content" }}>
-        <Tabs
-          activeKey={`${systemIndex}`}
-          onChange={(activeKey) =>
-            setActiveSystemExamples({
-              ...activeSystemExamples,
-              systemIndex: Number(activeKey),
-            })
-          }
-        >
-          {systems.map((system, sysIndex) => {
-            return (
-              <TabPane tab={system.system_info.system_name} key={`${sysIndex}`}>
-                <AnalysisTable
-                  systemID={system.system_id}
-                  task={task}
-                  cases={sortedBucketOfCasesList[sysIndex]}
-                  page={page}
-                  onPageChange={setPage}
-                />
-              </TabPane>
-            );
-          })}
-        </Tabs>
-      </Space>
-    );
-  }
-
-  const barText = systems.length === 1 ? "bar" : "bars";
-  const exampleText = "Examples";
-  exampleTable = (
-    <div>
-      <Title level={4}>{`${exampleText} from ${barText} #${
-        barIndex + 1
-      } in ${title}`}</Title>
-      {exampleTable}
-    </div>
-  );
-  return exampleTable;
 }
 
 export function AnalysisReport(props: Props) {
@@ -125,13 +41,16 @@ export function AnalysisReport(props: Props) {
     setPage(0);
   }
 
-  // Create the example table if a bar is selected, empty element if not
-  const exampleTable = createExampleTable(
-    props,
-    activeSystemExamples,
-    setActiveSystemExamples,
-    page,
-    setPage
+  /** Create the example table if a bar is selected, empty element if not */
+  const exampleTable = (
+    <ExampleTable
+      task={props.task}
+      systems={props.systems}
+      activeSystemExamples={activeSystemExamples}
+      setActiveSystemExamples={setActiveSystemExamples}
+      page={page}
+      setPage={setPage}
+    />
   );
 
   return (
